Extract Tab component to a new file
