+++ src/components/common/Tabs/Tab/index.tsx
@@ -1,0 +1,25 @@
+import { NavLink } from 'react-router-dom';
+import TabType from 'types/Tab';
+
+interface TabProps {
+  tab: TabType;
+}
+
+function Tab({ tab }: TabProps) {
+  return (
+    <li className="mr-1.5">
+      <NavLink
+        to={`.${tab.path}`}
+        className={({ isActive }) =>
+          `inline-block border-b-2 p-3 hover:border-gray-300 hover:text-gray-600 ${
+            isActive ? 'border-primary text-primary' : 'border-transparent'
+          }`
+        }
+      >
+        {tab.label}
+      </NavLink>
+    </li>
+  );
+}
+
+export default Tab;
